refactor: remove the state on Toolbar
