+++ src/index.ts
@@ -1,4 +1,4 @@
 import { RootLayout } from './layout/RootLayout';
-import { Toolbar } from './layout/SideBarLayout';
+import { Toolbar } from './layout/ToolbarLayout';
 
 export { RootLayout, Toolbar };
