--- stories/page.stories.tsx
@@ -1,17 +1,0 @@
-import { Meta } from '@storybook/react';
-import React from 'react';
-
-import App from '../src/App';
-
-export default {
-  title: 'Default',
-} as Meta;
-
-export function Control() {
-  return (
-    <>
-      <App />
-      <button type="button">Normalized button</button>
-    </>
-  );
-}
